don't depend on cstdint

+++ cmake/tests/cxx11_std_atomic.cpp
@@ -6,10 +6,9 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 #include <atomic>
-#include <cstdint>
 
 struct P {
-    uint64_t x; uint64_t y;
+    long long x; long long y;
 };
 
 int main()
